bills backend api and routing

+++ db/migrate/20200512110741_create_bills.rb
@@ -7,7 +7,7 @@ class CreateBills < ActiveRecord::Migration[6.0]
       t.references :user, type: :uuid, null: false, foreign_key: true
       t.references :patient, type: :uuid, null: false, foreign_key: true
       t.integer :bill_total, default: 0
-      t.decimal :paid_amount, default: 0
+      t.integer :paid_amount, default: 0
       t.boolean :payment_status, default: false
 
       t.timestamps
